refactoring config and key management
