+++ agent-pkg/src/index.ts
@@ -5,7 +5,9 @@ import { publish } from './commands/publish.js';
 import { list } from './commands/list.js';
 import { config } from './commands/config.js';
 import { clean } from './commands/clean.js';
-import { initializeConfigAndCredentials, credentialsExist } from './config/index.js';
+import { initializeConfigAndCredentials } from './config/index.js';
+import { credentialsExist } from './utils/credential/index.js';
+
 
 const checkMasterPassword = async () => {
   const isConfigCommand = process.argv.length > 2 && process.argv[2] === 'config';
